+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Bootstrap project
 - spring boot database connection configuration
 - Value Model
+- Value JPA Repository